feat: Make user tools easier to activate/deactivate
